Struct ddc_winapi::Monitor

pub struct Monitor { /* fields omitted */ }

A handle to an attached monitor that allows the use of DDC/CI operations.

Methods

impl Monitor

pub unsafe fn new(monitor: PHYSICAL_MONITOR) -> Self

Create a new monitor from the specified handle.

pub fn enumerate() -> Result<Vec<Self>>

Enumerate all connected physical monitors.

pub fn description(&self) -> String

Physical monitor description string.

pub fn handle(&self) -> HANDLE

Physical monitor winapi handle.

pub fn winapi_get_timing_report(&self) -> Result<MC_TIMING_REPORT>

Retrieves a monitor's horizontal and vertical synchronization frequencies.

pub fn winapi_set_vcp_feature(&self, code: BYTE, value: DWORD) -> Result<()>

Sets the value of a Virtual Control Panel (VCP) code for a monitor.

pub fn winapi_save_current_settings(&self) -> Result<()>

Saves the current monitor settings to the display's nonvolatile storage.

pub fn winapi_get_vcp_feature_and_vcp_feature_reply(
&self,
code: BYTE
) -> Result<(u32, DWORD, DWORD)>

Retrieves the current value, maximum value, and code type of a Virtual Control Panel (VCP) code for a monitor.

Returns (vcp_type, current_value, max_value)

pub fn winapi_get_capabilities_string_length(&self) -> Result<DWORD>

Retrieves the length of the buffer to pass to winapi_capabilities_request_and_capabilities_reply.

pub fn winapi_capabilities_request_and_capabilities_reply(
&self,
string: &mut [u8]
) -> Result<()>

Retrieves a string describing a monitor's capabilities.

This string is always ASCII and includes a terminating null character.